refactor(admin): key-management modals 폴더 → components로 이름 변경, ConfirmDialog 제거
