+++ apps/yongin-platform-admin/src/pages/key-management/components/index.ts
@@ -1,2 +1,1 @@
 export { KeyManagementModal } from "./KeyManagementModal";
-export { ConfirmDialog } from "./ConfirmDialog";
