--- apps/yongin-platform-admin/src/pages/key-management/modals/ConfirmDialog.tsx
@@ -1,50 +1,0 @@
-import { useState } from "react";
-import {
-  Modal,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalTitle,
-  ModalDescription,
-  Button,
-} from "@pf-dev/ui";
-import { ConfirmDialogProps } from "../types";
-
-export function ConfirmDialog({
-  open,
-  onOpenChange,
-  title,
-  description,
-  onConfirm,
-}: ConfirmDialogProps) {
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleConfirm = async () => {
-    try {
-      setIsLoading(true);
-      await onConfirm();
-      onOpenChange?.(false);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <Modal open={open} onOpenChange={onOpenChange}>
-      <ModalContent>
-        <ModalHeader>
-          <ModalTitle>{title}</ModalTitle>
-          <ModalDescription>{description}</ModalDescription>
-        </ModalHeader>
-        <ModalFooter>
-          <Button variant="outline" onClick={() => onOpenChange?.(false)} disabled={isLoading}>
-            취소
-          </Button>
-          <Button onClick={handleConfirm} disabled={isLoading}>
-            {isLoading ? "처리 중..." : "확인"}
-          </Button>
-        </ModalFooter>
-      </ModalContent>
-    </Modal>
-  );
-}
